+++ packages/react/CHANGELOG.md
@@ -3,6 +3,32 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.8.0-beta.4](https://github.com/DFXswiss/packages/compare/@dfx.swiss/react@1.8.0-beta.3...@dfx.swiss/react@1.8.0-beta.4) (2026-08-20)
+
+**Note:** Version bump only for package @dfx.swiss/react
+
+# [1.8.0-beta.3](https://github.com/DFXswiss/packages/compare/@dfx.swiss/react@1.8.0-beta.2...@dfx.swiss/react@1.8.0-beta.3) (2026-08-14)
+
+**Note:** Version bump only for package @dfx.swiss/react
+
+# [1.8.0-beta.2](https://github.com/DFXswiss/packages/compare/@dfx.swiss/react@1.8.0-beta.1...@dfx.swiss/react@1.8.0-beta.2) (2026-08-12)
+
+### Features
+
+- **core:** add the Yapeal personal IBAN provider member ([#211](https://github.com/DFXswiss/packages/issues/211)) ([2880e00](https://github.com/DFXswiss/packages/commit/2880e00f161f311a15bc6e454dc08937fa5c3754))
+
+# [1.8.0-beta.1](https://github.com/DFXswiss/packages/compare/@dfx.swiss/react@1.8.0-beta.0...@dfx.swiss/react@1.8.0-beta.1) (2026-08-11)
+
+### Bug Fixes
+
+- **react:** keep the support chat sync alive, and let a failed message be sent again ([#210](https://github.com/DFXswiss/packages/issues/210)) ([a3bace9](https://github.com/DFXswiss/packages/commit/a3bace9de2171ad179f7311028eedb7d7e38b6f7))
+
+# [1.8.0-beta.0](https://github.com/DFXswiss/packages/compare/@dfx.swiss/react@1.7.0-beta.1...@dfx.swiss/react@1.8.0-beta.0) (2026-08-07)
+
+### Features
+
+- **core,react:** let the buy invoice call target the collection account ([#204](https://github.com/DFXswiss/packages/issues/204)) ([d388d81](https://github.com/DFXswiss/packages/commit/d388d8105a539e56f78d246462c6ca021233bf39)), closes [DFXswiss/backend#4686](https://github.com/DFXswiss/backend/issues/4686) [DFXswiss/services#1270](https://github.com/DFXswiss/services/issues/1270) [#152](https://github.com/DFXswiss/packages/issues/152) [DFXswiss/backend#4686](https://github.com/DFXswiss/backend/issues/4686) [DFXswiss/services#1270](https://github.com/DFXswiss/services/issues/1270)
+
 # [1.7.0](https://github.com/DFXswiss/packages/compare/@dfx.swiss/react@1.7.0-beta.0...@dfx.swiss/react@1.7.0) (2026-07-29)
 
 ### Bug Fixes
@@ -43,7 +69,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ### Features
 
-- expose switchToCode on ApiException for merged-account redirects ([#187](https://github.com/DFXswiss/packages/issues/187)) ([2084e20](https://github.com/DFXswiss/packages/commit/2084e20c9ee3fd58e399d0a9a9e771f2fd135cc6)), closes [DFXswiss/api#4092](https://github.com/DFXswiss/api/issues/4092)
+- expose switchToCode on ApiException for merged-account redirects ([#187](https://github.com/DFXswiss/packages/issues/187)) ([2084e20](https://github.com/DFXswiss/packages/commit/2084e20c9ee3fd58e399d0a9a9e771f2fd135cc6)), closes [DFXswiss/backend#4092](https://github.com/DFXswiss/backend/issues/4092)
 
 ## [1.4.1](https://github.com/DFXswiss/packages/compare/@dfx.swiss/react@1.4.1-beta.0...@dfx.swiss/react@1.4.1) (2026-07-02)